% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{adjustSimMatrix}
\alias{adjustSimMatrix}
\title{A function that applies a weighting factor to the columns of a table}
\usage{
adjustSimMatrix(data, weights, default = 0)
}
\arguments{
\item{data}{a table with similarity scores}

\item{weights}{a vector with weights for the columns of the data table}

\item{default}{a default weight to be applied}
}
\description{
A function that applies a weighting factor to the columns of a table
}
