% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{calculateConnectionMatrix}
\alias{calculateConnectionMatrix}
\title{Calculates for each UID the connection as a list of two column results
on the first resulting column there is the first order type and on the 
second column is the subsequent order type.}
\usage{
calculateConnectionMatrix(orders, mapping)
}
\arguments{
\item{orders}{- a list of orders mapped per customer}

\item{mapping}{- a list of types per item of each order}
}
\description{
Calculates for each UID the connection as a list of two column results
on the first resulting column there is the first order type and on the 
second column is the subsequent order type.
}
