% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postRecommendations.R
\name{calculatePostOrderTable}
\alias{calculatePostOrderTable}
\title{Given a list of transactions the post order affinity table is calculated.
The expected format of transactions is a a data.table with
UID - unique customer identifier
reference - Grouping of orders
type - type per item in an order}
\usage{
calculatePostOrderTable(transactions)
}
\arguments{
\item{transactions}{- a data.table with transaction data}
}
\description{
Given a list of transactions the post order affinity table is calculated.
The expected format of transactions is a a data.table with
UID - unique customer identifier
reference - Grouping of orders
type - type per item in an order
}
