% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarProducts.R
\name{combineSimilarity}
\alias{combineSimilarity}
\title{Combined similarity to given products}
\usage{
combineSimilarity(sim.matrix, skus, exclude.same)
}
\arguments{
\item{sim.matrix}{similarity matrix based on implicit interractions}

\item{skus}{identifiers of the implicit interraction products (source should match names in sim.matrix)}

\item{exclude.same}{excludes values in `skus` from recommendations}
}
\description{
Turns recommendations matrix into a normalised data table
with mean similarity score for each recommended product
}
