% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{connectionsToCounts}
\alias{connectionsToCounts}
\title{Calculates the connections matrix given the pair of orders}
\usage{
connectionsToCounts(data)
}
\arguments{
\item{data}{- a data table with sku1 and sku2 columns that hold the first order type and second order type}
}
\description{
Calculates the connections matrix given the pair of orders
}
