% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{connectOrderMapping}
\alias{connectOrderMapping}
\title{The function that drills in each order and associates the types
based on the order history}
\usage{
connectOrderMapping(list.of.orders, order.mapping)
}
\arguments{
\item{list.of.orders}{- a single list of orders}

\item{order.mapping}{- a single list of items per order}
}
\description{
The function that drills in each order and associates the types
based on the order history
}
