% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosineCpp}
\alias{cosineCpp}
\title{Calculates cosine similarity matrix for a given matix}
\usage{
cosineCpp(x)
}
\arguments{
\item{x}{matrix of prouct hits}
}
\value{
matrix of product similarity
}
\description{
Tranforms product hits matrix to product similarity matrix
}
