% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{cosineMatrix}
\alias{cosineMatrix}
\title{Cosine similarity transformation for product hits}
\usage{
cosineMatrix(m)
}
\arguments{
\item{m}{matrix of product hit counts that happened in a single session}
}
\value{
product similarity matrix
}
\description{
Tranforms product hits matrix to product similarity matrix
}
