% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{excludeSame}
\alias{excludeSame}
\title{Filter out products that were seen by visitor}
\usage{
excludeSame(data)
}
\arguments{
\item{data}{sku combinations for prediction with sku.rec and sku fields}
}
\value{
sku combinations where sku.rec does not have seen sku within visitor.id
}
\description{
Filter out products that were seen by visitor
}
