% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{expandHits}
\alias{expandHits}
\title{Expand visitor product hits data to dataset for prediction}
\usage{
expandHits(object, data)
}
\arguments{
\item{object}{similarity model}

\item{data}{visitor page hits}
}
\description{
Expand visitor product hits data to dataset for prediction
}
