% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_concordance.R
\name{get_sample_concordance}
\alias{get_sample_concordance}
\title{Get Sample Concordance}
\usage{
get_sample_concordance(df, gs, meta, samples, sex)
}
\arguments{
\item{df}{A data frame containing gene expression data with 'gene_id' as
row names.}

\item{gs}{A data frame containing gene scores with at least two columns:
'gene_id' and 'score'.}

\item{meta}{A data frame containing metadata with at least three columns:
'mouse_id', 'sex', and 'strain'.}

\item{samples}{A vector of sample IDs to be included in the analysis.}

\item{sex}{A character string indicating the sex of the samples.
Default is 'both'.}
}
\value{
A data frame containing the strain, sex, Spearman correlation
coefficient, and the number of genes used in the calculation.
}
\description{
Calculates the Spearman correlation between gene expression data and gene
scores for each strain.
}
\details{
The function performs the following steps:
\itemize{
\item Filters the gene scores (\code{gs}) to include only those expressed in
the data (\code{df}).
\item Subsets the gene expression data (\code{df}) to include only the genes
present in the filtered gene scores.
\item Iterates over each unique strain in the metadata (\code{meta}).
\item For each strain, identifies the samples that match the given strain.
\item Calculates the mean expression for each gene across the identified samples
if there are multiple samples; otherwise, uses the single sample's expression.
\item Computes the Spearman correlation between the mean gene expression and the
gene scores.
}
}
\keyword{internal}
