% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementaryProducts.R
\name{getComplimentaryProducts}
\alias{getComplimentaryProducts}
\title{Function to return complimentary product recommendations}
\usage{
getComplimentaryProducts(sim.matrix, skus, values, exclude.same = TRUE,
  groups = NULL, affinity, sku.details, n.of.types)
}
\arguments{
\item{sim.matrix}{- similarity matrix.}

\item{skus}{- skus derived from the visitor history.}

\item{values}{- required number of recommendations.}

\item{exclude.same}{- excludes recommendations for values in skus.}

\item{groups}{- named vector of sku categories.}

\item{affinity}{- affinity matrix}

\item{sku.details}{- named vector of skus and types}

\item{n.of.types}{- number of types the results should be selected from}
}
\description{
Function to return complimentary product recommendations
}
