% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRecomendations.R
\name{getSimilarProducts}
\alias{getSimilarProducts}
\title{Provides similar product recommendations}
\usage{
getSimilarProducts(sim.matrix, skus, values, exclude.same, groups = NULL)
}
\arguments{
\item{sim.matrix}{similarity matrix}

\item{skus}{identifiers of the products that are used as names in sim.matrix}

\item{values}{required number of recommendations}

\item{exclude.same}{excludes recommendations for values in skus}

\item{groups}{- named vector of sku categories}
}
\description{
Provides similar product recommendations
}
