% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation-functions.R
\name{ib_collaborative_filtering}
\alias{ib_collaborative_filtering}
\title{ib_collaborative_filtering}
\usage{
ib_collaborative_filtering(df, n_recommendations = 10, n_neighbours = 5)
}
\arguments{
\item{df}{data.frame containing user-item matrix with grades that users give to items they know.}

\item{n_recommendations}{Number of recommendations to be returned for every user.}

\item{n_neighbours}{Number of item neighbours used when calculating grade predictions.}
}
\value{
matrix containing IDs of recommended items for each user. It has dimensions number-of-users x n_recommendations
}
\description{
Item-based collaborative filtering implementation. It searches for items similar to each other and from that it derives recommendations
(if user graded an item highly, it will probably grade items similar to it highly too).
}
