% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarProducts.R
\name{keepOnePerGroup}
\alias{keepOnePerGroup}
\title{Gets top value per group}
\usage{
keepOnePerGroup(dt, groups)
}
\arguments{
\item{dt}{data.table with sim}

\item{groups}{named vector of product groups}
}
\description{
Gets top value per group
}
