% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{makeRecommendationsFilter}
\alias{makeRecommendationsFilter}
\title{Create filter function to reduce number of recommendations
to a relevant subset}
\usage{
makeRecommendationsFilter(groups = NULL, values = 20)
}
\arguments{
\item{groups}{named vector of product types (or other level of product hierarchy)}

\item{values}{number of recommendations to return per visitor}
}
\description{
Create filter function to reduce number of recommendations
to a relevant subset
}
