% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarProducts.R
\name{notInWhich}
\alias{notInWhich}
\title{Creates permutation index for exclusion of values}
\usage{
notInWhich(x, y, filter)
}
\arguments{
\item{x}{vector to filter}

\item{y}{vector to match}

\item{filter}{flag to indicate whether filter should be applied}
}
\description{
Creates permutation index for exclusion of values
}
