% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{predict.similarity.recommender}
\alias{predict.similarity.recommender}
\title{Predicts similarity score for new product hits data}
\usage{
\method{predict}{similarity.recommender}(object, newdata)
}
\arguments{
\item{object}{similarity model object}

\item{newdata}{product hits data}
}
\description{
Predicts similarity score for new product hits data
}
