% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_gs.R
\name{preprocess_gs}
\alias{preprocess_gs}
\title{Preprocess Gene Scores}
\usage{
preprocess_gs(gs, verbose = TRUE)
}
\arguments{
\item{gs}{A data frame containing gene scores with at least two columns: 'gene_id' and 'score'.}

\item{verbose}{A logical value indicating whether to print detailed messages during processing. Default is TRUE.}
}
\value{
A cleaned data frame for the input geneset containing only 'gene_id' and 'score' columns of mouse Ensembl genes.
}
\description{
This function preprocesses a data frame of gene scores by performing several checks and cleaning steps.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if the input data frame contains the required columns 'gene_id' and 'score'.
\item Prints the number of unique genes in the gene set if \code{verbose} is TRUE.
\item Removes rows where 'gene_id' does not match the mouse Ensembl gene ID pattern.
\item Removes duplicate rows, keeping the score with the highest absolute value for duplicated 'gene_id's.
\item Checks if the 'score' column is numeric.
\item Prints a summary of the score range if \code{verbose} is TRUE.
\item Returns the cleaned data frame with only 'gene_id' and 'score' columns.
}
}
\examples{
\dontrun{
gs <- data.frame(gene_id = c("ENSMUSG00000000001", "ENSMUSG00000000003", "INVALID_ID"),
                 score = c(1.5, 2.3, 3.1))
cleaned_gs <- preprocess_gs(gs, verbose=TRUE)
}

}
\keyword{internal}
