% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{recommend}
\alias{recommend}
\title{Recommend Vulnerability Scores}
\usage{
recommend(
  gs,
  meta,
  df,
  seed = 303,
  B = 4000,
  combine_gs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gs}{A data frame containing gene scores with at least two columns: 'gene_id' and 'score'.
If a 'geneset' column is present and \code{combine=TRUE}, scores will be normalized within each
geneset before processing.}

\item{meta}{A data frame containing metadata with at least three columns: 'mouse_id', 'sex', and 'strain'.}

\item{df}{A data frame containing gene expression data with 'gene_id' as one of the columns.}

\item{seed}{An integer specifying the random seed for reproducibility.
Default is 303.}

\item{B}{An integer specifying the number of bootstrap samples for the null
distribution. Default is 4000.}

\item{combine_gs}{A logical value indicating if there are multiple gene sets to combine before running the function.
If so, gs data frame must contain genesets column to indicate different gene sets. Default is FALSE.}

\item{verbose}{A logical value indicating whether to print detailed messages. Default is TRUE.}
}
\value{
A data frame containing the vulnerability scores and false positive rates (FPR) for each sample/strain.
}
\description{
This function recommends vulnerability scores for each sample based on gene expression data and gene scores.
}
\details{
The function performs several checks and preprocessing steps:
\itemize{
\item Loads necessary libraries and sources required functions.
\item Preprocesses the gene set using the \code{preprocess_gs} function.
\item Ensures that \code{gs}, \code{meta}, and \code{df} are data frames.
\item Checks that \code{df} contains a 'gene_id' column with valid mouse Ensembl gene IDs.
\item Sets row names of \code{df} to 'gene_id' and removes the 'gene_id' column.
\item Ensures that all column names of \code{df} are present in \code{meta$mouse_id}.
\item Checks for the presence of 'sex' and 'strain' columns in \code{meta}.
\item Subsets \code{meta} to include only samples present in \code{df}.
}

The function then calculates vulnerability scores for each sample by:
\itemize{
\item Calculating concordance for each sample using the \code{get_sample_concordance} function.
\item Generating a null distribution of Spearman correlation coefficients using the \code{generate_concordance_null_distribution} function.
\item Calculating the false positive rate (FPR) for each vulnerability score.
}
}
\examples{
gs <- data.frame(gene_id = c("ENSMUSG00000000001", "ENSMUSG00000000003"),
                 score = c(1.5, 2.3))
meta <- data.frame(mouse_id = c("sample1", "sample2"),
                   sex = c("both", "both"),
                   strain = c("strain1", "strain2"))
df <- data.frame(gene_id = c("ENSMUSG00000000001", "ENSMUSG00000000003"),
                 sample1 = c(5.2, 3.8), sample2 = c(4.1, 2.9))
result <- recommend(gs, meta, df, verbose=TRUE)

}
