% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation-functions.R
\name{recommendation-functions}
\alias{recommendation-functions}
\alias{help.get_cosine}
\title{Recommendation functions.}
\usage{
help.get_cosine(x, y)
}
\description{
Recommendation functions.
}
