% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{recommendComplimentaryProducts}
\alias{recommendComplimentaryProducts}
\title{Recommend products in product-to-product scenario}
\usage{
recommendComplimentaryProducts(model, skus, group.column = "sku",
  limit = 20L)
}
\arguments{
\item{model}{similarity model object}

\item{skus}{data.table of product details with sku field as unique identifier.}

\item{group.column}{name of the column that will be used for grouping}

\item{limit}{number of records to return per product}
}
\description{
Recommend products in product-to-product scenario
}
