% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender-package.R
\name{recommender}
\alias{recommender}
\title{Simple recommendation library}
\description{
Library implementing collaborative filtering methods to make recommendations.
}
\details{
Contains two functions: \code{\link{ub_collaborative_filtering}} for user-based collaborative filtering
and \code{\link{ib_collaborative_filtering}} for item-based collaborative filtering.
}
