% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarProducts.R
\name{recommendSimilar}
\alias{recommendSimilar}
\title{Recommends products similar to given products}
\usage{
recommendSimilar(sim.matrix, skus, values, exclude.same, groups = NULL)
}
\arguments{
\item{sim.matrix}{similarity matrix based on implicit interractions}

\item{skus}{identifiers of the implicit interraction products (source should match names in sim.matrix)}

\item{values}{required number of recommendations}

\item{exclude.same}{excludes values in `skus` from recommendations}

\item{groups}{- named vector of sku categories}
}
\description{
Based on the similarity matix and given product recommends top products
}
