% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{recommendSimilarProducts}
\alias{recommendSimilarProducts}
\title{Recommend similar products to visitors based on product interraction}
\usage{
recommendSimilarProducts(model, hits, exclude.same = TRUE,
  filter = makeRecommendationsFilter())
}
\arguments{
\item{model}{similarity model object}

\item{hits}{visitor product hits data to be used for recommendations}

\item{exclude.same}{exludes products in the hits data per user if set to TRUE}

\item{filter}{function generated with makeRecommendationsFilter()}
}
\description{
Recommend similar products to visitors based on product interraction
}
