% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{similarity.predictor}
\alias{similarity.predictor}
\title{Predicts similarity score for new product hits data}
\usage{
similarity.predictor(object, newdata)
}
\arguments{
\item{object}{similarity model object}

\item{newdata}{product hits data}
}
\description{
Predicts similarity score for new product hits data
}
