% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\docType{methods}
\name{predict,similarity.recommender-method}
\alias{predict,similarity.recommender-method}
\alias{predict,similarity-recommender}
\title{Makes similarity score predictions based on the similarity.recommender model}
\usage{
\S4method{predict}{similarity.recommender}(object, newdata)
}
\arguments{
\item{object}{a fitted similarity model object.}

\item{newdata}{use product hits.}
}
\value{
\code{predict} returns a vector with predicted values.
}
\description{
Makes similarity score predictions based on the similarity.recommender model
}
