% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\docType{class}
\name{similarity.recommender-class}
\alias{similarity.recommender-class}
\title{S4 class that represents similar products recommendation model}
\arguments{
\item{sim}{cosine matrix of product similarity}
}
\description{
S4 class that represents similar products recommendation model
}
