% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarModel.R
\name{similarityRecommender}
\alias{similarityRecommender}
\title{Factory for the similarity recommendation model}
\usage{
similarityRecommender(data, filter = NULL)
}
\arguments{
\item{data}{product hit stream with visitor.id and sku}

\item{filter}{allows to reduce recommendations to a given set}
}
\description{
Factory for the similarity recommendation model
}
