% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{simplify.transactions}
\alias{simplify.transactions}
\title{Simplify two vector lists}
\usage{
simplify.transactions(f, s)
}
\arguments{
\item{f}{- Vector to be used as the grouping}

\item{s}{- Vector to include the values per grouping category}
}
\description{
Simplify two vector lists
}
