% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation-functions.R
\name{ub_collaborative_filtering}
\alias{ub_collaborative_filtering}
\title{ub_collaborative_filtering}
\usage{
ub_collaborative_filtering(df, n_recommendations = 10, n_neighbours = 5)
}
\arguments{
\item{df}{data.frame containing user-item matrix with grades that users give to items they know.}

\item{n_recommendations}{Number of recommendations to be returned for every user.}

\item{n_neighbours}{Number of user neighbours used when calculating grade predictions.}
}
\value{
matrix containing IDs of recommended items for each user. It has dimensions number-of-users x n_recommendations
}
\description{
User-based collaborative filtering implementation. For every user it searches its nearest neighbours
(users most similar to him) and derives grade predictions from grades that his neighbours gave to items he didn't grade himself. Finally
for every user it returns a list of recommendations - items with highest predicted grade.
}
