% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{userProductHitsToMatrix}
\alias{userProductHitsToMatrix}
\title{Turns interactions data table into a matrix}
\usage{
userProductHitsToMatrix(product.hits)
}
\arguments{
\item{product.hits}{data.table with columns:  visitor.id, product attribute}
}
\value{
Returns sparse matrix for interactions between products within user sessions,
  meaning number of visitors that viewed/bought products together.
}
\description{
Turns interactions data table into a matrix
}
