% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemRec.R
\name{recommend}
\alias{recommend}
\title{Recommender for books
The package recommends a list of books according to user, or item.}
\usage{
recommend(data, criteria, type = "", raw_data = NULL)
}
\arguments{
\item{data}{database of book and user ratings}

\item{criteria}{user wanting the recommended books}

\item{typetype}{of recommeder system}
}
\value{
the top 10 books like to be liked by user.
}
\description{
Recommender for books
The package recommends a list of books according to user, or item.
}
\examples{
recommend(book_ratings, 13, type = 'item')
recommend(book_ratings, 23, type = 'user')
}
